requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-google-genai>=0.3.0
langchain-core>=0.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0


sqlite3


requests>=2.28.0


pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0